DPC-678: Moved resource tests into v1 directory.
